Avoid pinning dependencies unless absolutely required
We had started pinning our requirements which led to an issue where a
library we pinned (urllib3) conflicted with a version needed by the AI
Assistant (which uses this SDK).  Since we actually don't care about the
version of that library, we really don't need to pin it in our project
dependencies (pinning in our requirements file is fine and, in fact,
preferred).

Fixes #26