Add support for targetNamespace to subscribe
Also added the ability to install a handler which is called when the
subscriber's connection is terminated.  Improved cleanup in this case to
make sure subscriber is reset to "disconnected" state.

Added support for protocol being added in 1.39 to indicate when the
server side removes a subscription (typically due to the removal of a
namespace).

Fixes #38